hold

+++ account-json/src/main/java/com/qa/accapp/App.java
@@ -12,7 +12,7 @@ public class App {
         list.addToList(Oli);
         String list2 = list.convertToJson(list);
         System.out.println(list2);
-        int num = list.count("Jack");
+        long num = list.count("Jack");
         System.out.println(num);
     }
 }
